+++ resources/leiningen/new/luminus/reframe/src/cljs/events.cljs
@@ -1,0 +1,32 @@
+(ns <<project-ns>>.events
+  (:require [<<project-ns>>.db :as db]
+            [re-frame.core :refer [dispatch reg-event-db reg-sub]]))
+
+;;dispatchers
+
+(reg-event-db
+  :initialize-db
+  (fn [_ _]
+    db/default-db))
+
+(reg-event-db
+  :set-active-page
+  (fn [db [_ page]]
+    (assoc db :page page)))
+
+(reg-event-db
+  :set-docs
+  (fn [db [_ docs]]
+    (assoc db :docs docs)))
+
+;;subscriptions
+
+(reg-sub
+  :page
+  (fn [db _]
+    (:page db)))
+
+(reg-sub
+  :docs
+  (fn [db _]
+    (:docs db)))